+++ packages/utils/src/client/index.ts
@@ -1,2 +1,4 @@
+export * from './clipboard';
 export * from './downloadFile';
 export * from './exportFile';
+export * from './sanitize';